refactor(roadmaps): extract services, add utils and documentation
- move roadmap logic into dedicated services under `services/`.

- update `RoadmapsModule` imports to use new service paths.

- add `README.md` documenting backend architecture.

- introduce utility helpers for async, date, markdown, milestone output,
  number, quiz, access, formatters, generation, records, constants, and
  timeline.

- add comprehensive unit tests for utils and service facade.

- adjust existing imports and remove unused code in
  `roadmaps.service.ts`.

- update Dagre layout import path and related tests.
