+++ apps/api/src/modules/roadmaps/utils/async.ts
@@ -1,0 +1,4 @@
+export const delay = (ms: number): Promise<void> =>
+  new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
