+++ apps/api/src/modules/roadmaps/utils/markdown.ts
@@ -1,0 +1,10 @@
+export const stripMarkdownFences = (text: string): string => {
+  const trimmed = text.trim();
+  if (trimmed.startsWith('```')) {
+    return trimmed
+      .replace(/^```(?:json)?/i, '')
+      .replace(/```$/, '')
+      .trim();
+  }
+  return trimmed;
+};
